# om-backend/BE_OpenAPI.yaml
openapi: 3.0.0
info:
  title: SMAUG OpenAPI
  description: SMAUG OpenAPI for the smart locker-related operations.
  version: 1.0.0
  contact:
    name: Antonio Antonino
    email: antonio.antonino@ericsson.com
...
servers:
  - url: /api/v1
paths:
...
  /marketplace/requests/{request_id}/offers:
    parameters:
      - in: path
        name: request_id
        description: The ID of the request to retrieve offers for.
        schema:
          type: integer
          minimum: 1
          maximum: 1.844674407E19
        required: true  
    get:
      operationId: getRequestOffers
      summary: "!!! PARTIALLY IMPLEMENTED !!! Retrieve the offers submitted for a request, optionally filtered by user's input. At the moment, user's filtering is not supported."
      tags:
        - Marketplace
      parameters:
        - in: query
          name: valid_at
          description: OPTIONAL. The time at which the offer must be valid. If not specified, all the offers are returned. **Currently not supported**.
          schema:
            type: string
            format: date-time
        - in: query
          name: valid_for
          description: OPTIONAL. How long must the offer be valid, in minutes, from the time specified in `valid_at` (e.g. 30 min). If `valid_at` is specified, it must also be specified. **Currently not supported**.
          schema:
            type: integer
            minimum: 1
            maximum: 1.844674407E19
        - in: query
          name: only_auction
          description: OPTIONAL, default false. If true, restricts the results to only contain auction offers, with no instant rent. **Currently not supported**.
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: Returns all the offer IDs matching the user's requirements for the given request.
          content:
            application/json:
              schema:
                type: array
                description: The list of offers for request.
                items:
                  type: integer
                  minimum: 1
                  maximum: 1.844674407E19
              example:
                  - 2
                  - 3
                  - 14
        400:
          description: Error in the query parameters.
        404:
          description: Request with given ID not found.
        5XX:
          description: Unexpected error.   
  /marketplace/gettoken:
    get:
      operationId: getMarketplaceToken
      summary: Return a token required to create a new request on the blockchain.
      tags:
        - Marketplace
      security:
        - BearerAuth: []  
      parameters:
        - in: query
          name: ethereum_address
          description: The Ethereum account that will create the request.
          schema:
            type: string
          required: true          
      responses:
        201:
          description: Returns the access token.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/MarketplaceAccessToken'
                example:
                  digest: '0xc4dabc49eb44973021d2f4ef1681fde9b74099c168d54f764bd3040c820dea3d'
                  encoded: '0xa1f7afe9841e508b1baf684389c2cc6032188a202e6cfafbd02187c4e91b30ad6eb99a1a471e0575bFC76d7e189ab3354E0ecb70FCbf3E4651Cdd045E893528fB7C6F2DCEB07Efa9eb2FB375'
                  signature: '0x2e6e106429ce2d1c1519db56a8411bfa421048282f692b8604a1e51688319f4c25fd88833bafeccd4b89f04c65965966943a0256a1f18847900148585a8086fc1b'
                  nonce: '0xa1f7afe9841e508b1baf684389c2cc6032188a202e6cfafbd02187c4e91b30ad'
        400:
          description: Invalid Ethereum address parameter.
        422:
          description: Missing required query parameter.
        5XX:
          description: Unexpected error.
...
  /registration:
    post:
      operationId: userRegistration
      summary: "Register a new user with DID."
      tags:
        - Registration
      requestBody:
        description: The user registers by providing a fully-qualified DID and the related verification key (verkey).
        required: true
        content:
          application/json:
            schema:
              type: object
            example:
              did: "did:sov:V7tqpU5b3khe88hM6gehBb"
              verkey: "GKtmbQEQ8M8CpGdEAQ3Kv75buqcUsQaLM5F4j3i1awVD"
      responses:
        201:
          description: User succesfully registered.
        422:
          description: Bad data in input.
        5XX:
          description: Unexpected error.
...
  /gettoken:
    post:
      operationId: getToken
      summary: "Solve a challenge and Get the token to authenticate in the system."
      tags:
        - Authentication
      requestBody:
        description: The user return the solved challenge to the system to prove ownership of a given DID.
        required: true
        content:
          application/json:
            schema:
              type: object
            example:
              did: "did:sov:V7tqpU5b3khe88hM6gehBb"
              challenge: "001584719795427360495340695794"
              proof: "YWRzYWRzYXFlZm5qaWZ3cWVuYmpmbndsZ2Z3cWZnbnF3Z2xxd2dn"
      responses:
        200:
          description: Challenge response is valid and a JWT has been generated.
        422:
          description: Bad data in input.
        401:
          description: Proof not valid.
        403:
          description: Invalid or missing input parameters.
        5XX:
          description: Unexpected error.
components:
  schemas:
...
    MarketplaceAccessToken:
      description: Access token used to create new requests on the marketplace.
      properties:
        digest:
          description: The SHA-3 hash of the access token, in HEX format.
          type: string
        encoded:
          description: An Ethereum ABI encoding of the information in the access token.
          type: string
        signature:
          description: The signature of the access token.
          type: string
        nonce:
          description: The randomly-generated nonce for the access token, in HEX format.
          type: string
          pattern: '^[0-9a-f]*$'
      required:
        - digest
        - encoded
        - signature
        - nonce
...
  securitySchemes:
    BearerAuth:
      description: Protects endpoints by requiring a valid JWT in the request `Authorization` header.
      type: http
      scheme: bearer
      bearerFormat: JWT